+++ src/modules/dashboard/sagas/motions/index.ts
@@ -1,0 +1,7 @@
+import { all, call } from 'redux-saga/effects';
+
+import stakeMotionSaga from './stakeMotion';
+
+export default function* actionsSagas() {
+  yield all([call(stakeMotionSaga)]);
+}
