+++ src/modules/dashboard/sagas/motions/stakeMotion.ts
@@ -1,100 +1,18 @@
 import { call, fork, put, takeEvery } from 'redux-saga/effects';
 import { ClientType, ExtensionClient } from '@colony/colony-js';
-import { BigNumber } from 'ethers/utils';
 
 import { Action, ActionTypes, AllActions } from '~redux/index';
 import { TEMP_getContext, ContextModule } from '~context/index';
 import { putError, takeFrom } from '~utils/saga/effects';
-import { Address } from '~types/index';
-import {
-  MotionStakesQuery,
-  MotionStakesQueryVariables,
-  MotionStakesDocument,
-  EventsForMotionQuery,
-  EventsForMotionQueryVariables,
-  EventsForMotionDocument,
-  MotionsSystemMessagesQuery,
-  MotionsSystemMessagesQueryVariables,
-  MotionsSystemMessagesDocument,
-  ColonyActionQuery,
-  ColonyActionQueryVariables,
-  ColonyActionDocument,
-} from '~data/index';
 
 import {
   createTransaction,
   createTransactionChannels,
   getTxChannel,
-} from '../../core/sagas';
+} from '../../../core/sagas';
+import { transactionReady } from '../../../core/actionCreators';
 
-import { transactionReady } from '../../core/actionCreators';
-
-export function* updateCacheValues(
-  colonyAddress: Address,
-  userAddress: Address,
-  motionId: BigNumber,
-  transactionHash: string,
-) {
-  const apolloClient = TEMP_getContext(ContextModule.ApolloClient);
-
-  /*
-   * Staking values
-   */
-  yield apolloClient.query<MotionStakesQuery, MotionStakesQueryVariables>({
-    query: MotionStakesDocument,
-    variables: {
-      colonyAddress,
-      userAddress,
-      motionId: motionId.toNumber(),
-    },
-    fetchPolicy: 'network-only',
-  });
-
-  /*
-   * Motion Events
-   */
-  yield apolloClient.query<EventsForMotionQuery, EventsForMotionQueryVariables>(
-    {
-      query: EventsForMotionDocument,
-      variables: {
-        colonyAddress,
-        motionId: motionId.toNumber(),
-      },
-      fetchPolicy: 'network-only',
-    },
-  );
-
-  /*
-   * Motion System Messages
-   */
-  yield apolloClient.query<
-    MotionsSystemMessagesQuery,
-    MotionsSystemMessagesQueryVariables
-  >({
-    query: MotionsSystemMessagesDocument,
-    variables: {
-      colonyAddress,
-      motionId: motionId.toNumber(),
-    },
-    fetchPolicy: 'network-only',
-  });
-
-  /*
-   * Colony Actions (to get the refreshed motion state)
-   *
-   * @NOTE It might make sense in the long run to just create a separate
-   * resolver just for the motion's state. It will cut down on fetching
-   * data we don't need just to show the updated state
-   */
-  yield apolloClient.query<ColonyActionQuery, ColonyActionQueryVariables>({
-    query: ColonyActionDocument,
-    variables: {
-      colonyAddress,
-      transactionHash,
-    },
-    fetchPolicy: 'network-only',
-  });
-}
+import { updateMotionValues } from '../utils/updateMotionValues';
 
 function* stakeMotion({
   meta,
@@ -196,7 +114,7 @@ function* stakeMotion({
      * Update motion page values
      */
     yield fork(
-      updateCacheValues,
+      updateMotionValues,
       colonyAddress,
       userAddress,
       motionId,
@@ -215,6 +133,6 @@ function* stakeMotion({
   return null;
 }
 
-export default function* motionSagas() {
+export default function* stakeMotionSaga() {
   yield takeEvery(ActionTypes.COLONY_MOTION_STAKE, stakeMotion);
 }
